refactor(jxa): inline shared buildTask helper via @inline directive
Land ADR-0020 specifying a `// @inline <relative-path>` build-time
directive expanded by scriptInlinerPlugin (esbuild + Vite). Extend the
loader with `expandInlineDirectives` so consumer scripts can splice
shared helper sources at load time. Helpers live under
`src/scripts/jxa/_helpers/`.

Extract the canonical buildTask + buildRepetition into
`_helpers/build_task.js`. Reconciled across the 8 prior copies preserving
every issue-referenced fix verbatim (#673 containingProject().class()
quirk, #682 per-element tag-id guard, #498 creationDate() invocation
guard). The intentional semantic split between forecast/search and the
rest (task.effectivelyAvailable() vs task.available()) is preserved via
an `options.effectiveAvailability` flag.

Wire 8 task-side consumers via `// @inline _helpers/build_task.js`:
task_get, task_list, task_create, task_update, task_search,
task_get_many, forecast_get, perspective_evaluate.

Bump bundle budget 740 → 760 KiB to absorb the canonical helper's
docblock inlined verbatim into 8 consumers; net source diff is −1482
lines on consumers. Drift bugs of #673's class become structurally
impossible — every fix lands in one place.

Closes #686. Follow-ups: #703 (buildProject), #704 (buildTag), #705
(buildFolder) extract the remaining duplicated helpers using the same
mechanism.
