+++ DESIGN.md
@@ -836,7 +836,7 @@ Coverage is not enforced because it's gameable. Test *fidelity* is enforced at r
 - `pnpm lint` — zero errors; biome config enforces `coding.md` standards
 - `pnpm test` — zero failures; execution < 10s
 - `pnpm build` — single-file bundle emitted to `dist/index.js`
-- Bundle size budget: < 680 KiB (tsup --minify); above that blocks release. Bumped 500 → 525 → 540 → 580 → 610 → 625 → 640 → 660 → 680 KiB as the tool surface grew — per-tool string and Zod-schema overhead became the dominant bundle cost. The 660 → 680 bump landed alongside slice 1 of [#483](https://github.com/torsday/omnifocus-mcp/issues/483) (webhooks: registry + register/list/delete tools + types + capability-resource integration + env-flag wiring per ADR-0016). The 640 → 660 bump landed alongside slice 1 of [#485](https://github.com/torsday/omnifocus-mcp/issues/485) (decision-journal: decision_record + decision_clear tools, parser, read-side integration on get/get_many for both tasks and projects). The 625 → 640 bump landed alongside the final slice of [#484](https://github.com/torsday/omnifocus-mcp/issues/484) (calendar + agenda — bridge wrapper, calendar resource, agenda merge module, AgendaItem union, capabilities probe). The 610 → 625 bump landed alongside [#577](https://github.com/torsday/omnifocus-mcp/issues/577) (perspective_create + perspective_update added two OmniJS scripts inlined verbatim plus a recursive input rule schema). The 580 → 610 bump landed alongside [#570](https://github.com/torsday/omnifocus-mcp/issues/570) (Example: sweep added ~7 KiB of description strings). Further bumps should NOT be flat increases: [#578](https://github.com/torsday/omnifocus-mcp/issues/578) tracks the tree-shaking / code-splitting investigation that should replace the next bump.
+- Bundle size budget: < 760 KiB (tsup --minify); above that blocks release. Bumped 500 → 525 → 540 → 580 → 610 → 625 → 640 → 660 → 680 → 700 → 740 → 760 KiB as the tool surface grew — per-tool string and Zod-schema overhead, plus inlined script payload, are the dominant bundle costs. The 740 → 760 KiB bump landed alongside [#686](https://github.com/torsday/omnifocus-mcp/issues/686) (DRY JXA scripts via build-time helper inlining per ADR-0020 — net source decrease of −1482 lines, but the canonical helper's docblock is inlined verbatim into 8 task-side consumers). The 700 → 740 KiB bump landed alongside [#689](https://github.com/torsday/omnifocus-mcp/issues/689) (ban-empty-catch annotated ~280 catch blocks across 32 JXA/OmniJS scripts). The 680 → 700 KiB bump landed alongside [#681](https://github.com/torsday/omnifocus-mcp/issues/681). The 660 → 680 KiB bump landed alongside slice 1 of [#483](https://github.com/torsday/omnifocus-mcp/issues/483) (webhooks: registry + register/list/delete tools + types + capability-resource integration + env-flag wiring per ADR-0016). The 640 → 660 KiB bump landed alongside slice 1 of [#485](https://github.com/torsday/omnifocus-mcp/issues/485) (decision-journal: decision_record + decision_clear tools, parser, read-side integration on get/get_many for both tasks and projects). The 625 → 640 KiB bump landed alongside the final slice of [#484](https://github.com/torsday/omnifocus-mcp/issues/484) (calendar + agenda — bridge wrapper, calendar resource, agenda merge module, AgendaItem union, capabilities probe). The 610 → 625 KiB bump landed alongside [#577](https://github.com/torsday/omnifocus-mcp/issues/577) (perspective_create + perspective_update added two OmniJS scripts inlined verbatim plus a recursive input rule schema). The 580 → 610 KiB bump landed alongside [#570](https://github.com/torsday/omnifocus-mcp/issues/570) (Example: sweep added ~7 KiB of description strings). Further bumps should NOT be flat increases: [#578](https://github.com/torsday/omnifocus-mcp/issues/578) tracks the tree-shaking / code-splitting investigation that should replace the next bump.
 
 ### Required status checks (branch protection on `main`)
