+++ src/scripts/jxa/forecast_get.js
@@ -33,181 +33,7 @@ function run(argv) {
   const includeDeferred = args.includeDeferred !== false;
   const includeFlagged = args.includeFlagged !== false;
 
-  // ---------------------------------------------------------------------------
-  // buildTask — same shape as task_list.js / task_search.js
-  // ---------------------------------------------------------------------------
-
-  function buildRepetition(task) {
-    try {
-      const rr = task.repetitionRule();
-      if (!rr) return null;
-      return { method: rr.method(), unit: rr.unit(), steps: rr.steps() };
-    } catch (_e) {
-      return null;
-    }
-  }
-
-  function buildTask(task) {
-    // See task_get.js — `cp.class()` throws on real projects in OF 4.x (#673).
-    let projectId = null;
-    try {
-      const cp = task.containingProject();
-      if (cp) {
-        let isDocument = false;
-        try {
-          isDocument = cp.class() === "document";
-        } catch (_classErr) {
-          /* OF 4.x: real projects throw here */
-        }
-        if (!isDocument) projectId = cp.id();
-      }
-    } catch (_e) {
-      /* OF 4.x: property access may not exist on all object types — default used */
-    }
-
-    let parentId = null;
-    try {
-      const pt = task.parentTask();
-      if (pt) parentId = pt.id();
-    } catch (_e) {
-      /* OF 4.x: property access may not exist on all object types — default used */
-    }
-
-    const tagIds = [];
-    try {
-      const tags = task.tags();
-      for (let i = 0; i < tags.length; i++) {
-        try {
-          tagIds.push(tags[i].id());
-        } catch (_tagErr) {
-          /* OF 4.x: individual tag specifier may throw on .id() — skip element */
-        }
-      }
-    } catch (_e) {
-      /* OF 4.x: property access may not exist on all object types — default used */
-    }
-
-    let deferDate = null;
-    try {
-      const dd = task.deferDate();
-      if (dd) deferDate = dd.toISOString();
-    } catch (_e) {
-      /* OF 4.x: property access may not exist on all object types — default used */
-    }
-
-    let dueDate = null;
-    try {
-      const due = task.dueDate();
-      if (due) dueDate = due.toISOString();
-    } catch (_e) {
-      /* OF 4.x: property access may not exist on all object types — default used */
-    }
-
-    let completedAt = null;
-    try {
-      const cd = task.completionDate();
-      if (cd) completedAt = cd.toISOString();
-    } catch (_e) {
-      /* OF 4.x: property access may not exist on all object types — default used */
-    }
-
-    let dropped = false;
-    try {
-      dropped = task.dropped();
-    } catch (_e) {
-      /* OF 4.x: property access may not exist on all object types — default used */
-    }
-
-    let flagged = false;
-    try {
-      flagged = task.flagged();
-    } catch (_e) {
-      /* OF 4.x: property access may not exist on all object types — default used */
-    }
-
-    let isCompleted = false;
-    try {
-      isCompleted = task.completed();
-    } catch (_e) {
-      /* OF 4.x: property access may not exist on all object types — default used */
-    }
-
-    let available = false;
-    try {
-      available = task.effectivelyAvailable ? task.effectivelyAvailable() : false;
-    } catch (_e) {
-      /* OF 4.x: property access may not exist on all object types — default used */
-    }
-
-    let blocked = false;
-    try {
-      blocked = task.blocked ? task.blocked() : false;
-    } catch (_e) {
-      /* OF 4.x: property access may not exist on all object types — default used */
-    }
-
-    let sequential = false;
-    try {
-      sequential = task.sequential ? task.sequential() : false;
-    } catch (_e) {
-      /* OF 4.x: property access may not exist on all object types — default used */
-    }
-
-    let completedByChildren = false;
-    try {
-      completedByChildren = task.completedByChildren ? task.completedByChildren() : false;
-    } catch (_e) {
-      /* OF 4.x: property access may not exist on all object types — default used */
-    }
-
-    let note = null;
-    try {
-      const raw = task.note ? task.note() : "";
-      note = raw || null;
-    } catch (_e) {
-      /* OF 4.x: property access may not exist on all object types — default used */
-    }
-
-    return {
-      id: task.id(),
-      name: task.name(),
-      note,
-      noteHtml: null,
-      projectId,
-      parentId,
-      tagIds,
-      flagged,
-      dueDate,
-      deferDate,
-      completedAt,
-      droppedAt: null,
-      completed: isCompleted,
-      dropped,
-      available,
-      blocked,
-      sequential,
-      completedByChildren,
-      estimatedMinutes: null,
-      repetition: buildRepetition(task),
-      // Guard against "Can't get object." thrown when invoking these — see #498.
-
-      createdAt: (() => {
-        try {
-          return task.creationDate().toISOString();
-        } catch (_e) {
-          return new Date().toISOString();
-        }
-      })(),
-
-      modifiedAt: (() => {
-        try {
-          return task.modificationDate().toISOString();
-        } catch (_e) {
-          return new Date().toISOString();
-        }
-      })(),
-    };
-  }
+  // @inline _helpers/build_task.js
 
   // ---------------------------------------------------------------------------
   // Collect candidates and bucket them — see #500.
@@ -244,7 +70,7 @@ function run(argv) {
   function builtFor(task) {
     const id = task.id();
     if (builtById[id] !== undefined) return builtById[id];
-    const b = buildTask(task);
+    const b = buildTask(task, { effectiveAvailability: true });
     builtById[id] = b;
     return b;
   }
