+++ src/scripts/jxa/task_create.js
@@ -19,197 +19,7 @@ function run(argv) {
   const ofApp = Application("OmniFocus");
   ofApp.includeStandardAdditions = false;
 
-  function buildRepetition(task) {
-    try {
-      const rr = task.repetitionRule();
-      if (!rr) return null;
-      return {
-        method: rr.method(),
-        unit: rr.unit(),
-        steps: rr.steps(),
-      };
-    } catch (_e) {
-      return null;
-    }
-  }
-
-  function buildTask(task) {
-    // See task_get.js — `cp.class()` throws on real projects in OF 4.x
-    // (#673). Treat a throw as "real project", a successful return as
-    // "document — skip".
-    let projectId = null;
-    try {
-      const cp = task.containingProject();
-      if (cp) {
-        let isDocument = false;
-        try {
-          isDocument = cp.class() === "document";
-        } catch (_classErr) {
-          /* OF 4.x: real projects throw here */
-        }
-        if (!isDocument) projectId = cp.id();
-      }
-    } catch (_e) {
-      /* OF 4.x: property access may not exist on all object types — default used */
-    }
-
-    let parentId = null;
-    try {
-      const pt = task.parentTask();
-      if (pt) parentId = pt.id();
-    } catch (_e) {
-      /* OF 4.x: property access may not exist on all object types — default used */
-    }
-
-    const tagIds = [];
-    try {
-      const tags = task.tags();
-      for (let i = 0; i < tags.length; i++) {
-        try {
-          tagIds.push(tags[i].id());
-        } catch (_tagErr) {
-          /* OF 4.x: individual tag specifier may throw on .id() — skip element */
-        }
-      }
-    } catch (_e) {
-      /* OF 4.x: property access may not exist on all object types — default used */
-    }
-
-    let deferDate = null;
-    try {
-      const dd = task.deferDate();
-      if (dd) deferDate = dd.toISOString();
-    } catch (_e) {
-      /* OF 4.x: property access may not exist on all object types — default used */
-    }
-
-    let dueDate = null;
-    try {
-      const due = task.dueDate();
-      if (due) dueDate = due.toISOString();
-    } catch (_e) {
-      /* OF 4.x: property access may not exist on all object types — default used */
-    }
-
-    let completedAt = null;
-    try {
-      const cd = task.completionDate();
-      if (cd) completedAt = cd.toISOString();
-    } catch (_e) {
-      /* OF 4.x: property access may not exist on all object types — default used */
-    }
-
-    let dropped = false;
-    try {
-      dropped = task.dropped();
-    } catch (_e) {
-      /* OF 4.x: property access may not exist on all object types — default used */
-    }
-
-    let estimatedMinutes = null;
-    try {
-      const em = task.estimatedMinutes();
-      if (em != null) estimatedMinutes = em;
-    } catch (_e) {
-      /* OF 4.x: property access may not exist on all object types — default used */
-    }
-
-    let note = null;
-    try {
-      note = task.note() || null;
-    } catch (_e) {
-      /* OF 4.x: property access may not exist on all object types — default used */
-    }
-
-    let noteHtml = null;
-    try {
-      if (task.noteHtml) noteHtml = task.noteHtml() || null;
-    } catch (_e) {
-      /* OF 4.x: property access may not exist on all object types — default used */
-    }
-
-    let flagged = false;
-    try {
-      flagged = task.flagged();
-    } catch (_e) {
-      /* OF 4.x: property access may not exist on all object types — default used */
-    }
-
-    let completed = false;
-    try {
-      completed = task.completed();
-    } catch (_e) {
-      /* OF 4.x: property access may not exist on all object types — default used */
-    }
-
-    let sequential = false;
-    try {
-      sequential = task.sequential();
-    } catch (_e) {
-      /* OF 4.x: property access may not exist on all object types — default used */
-    }
-
-    let completedByChildren = false;
-    try {
-      completedByChildren = task.containsSingletonActions();
-    } catch (_e) {
-      /* OF 4.x: property access may not exist on all object types — default used */
-    }
-
-    let available = false;
-    try {
-      available = task.available();
-    } catch (_e) {
-      /* OF 4.x: property access may not exist on all object types — default used */
-    }
-
-    let blocked = false;
-    try {
-      blocked = task.blocked();
-    } catch (_e) {
-      /* OF 4.x: property access may not exist on all object types — default used */
-    }
-
-    return {
-      id: task.id(),
-      name: task.name(),
-      note: note,
-      noteHtml: noteHtml,
-      projectId: projectId,
-      parentId: parentId,
-      tagIds: tagIds,
-      deferDate: deferDate,
-      dueDate: dueDate,
-      estimatedMinutes: estimatedMinutes,
-      flagged: flagged,
-      completed: completed,
-      completedAt: completedAt,
-      dropped: dropped,
-      droppedAt: null,
-      available: available,
-      blocked: blocked,
-      sequential: sequential,
-      completedByChildren: completedByChildren,
-      repetition: buildRepetition(task),
-      // Guard against "Can't get object." thrown when invoking these — see #498.
-
-      createdAt: (() => {
-        try {
-          return task.creationDate().toISOString();
-        } catch (_e) {
-          return new Date().toISOString();
-        }
-      })(),
-
-      modifiedAt: (() => {
-        try {
-          return task.modificationDate().toISOString();
-        } catch (_e) {
-          return new Date().toISOString();
-        }
-      })(),
-    };
-  }
+  // @inline _helpers/build_task.js
 
   if (!args.name || args.name.trim() === "") {
     throw new Error("ValidationError: name is required and cannot be empty");
